requirements.txt
asgiref==3.3.4
autopep8==1.5.7
certifi==2021.5.30
chardet==4.0.0
future==0.18.2
httpie==2.4.0
idna==2.10
iso8601==0.1.14
keyboard==0.13.5
pkg-resources==0.0.0
pycodestyle==2.7.0
Pygments==2.9.0
pyserial==3.5
PySocks==1.7.1
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
requests-toolbelt==0.9.1
RPi.GPIO==0.7.0
serial==0.0.97
sqlparse==0.4.1
subprocess.run==0.0.8
toml==0.10.2
typing-extensions==3.10.0.0
urllib3==1.26.5
usb-barcode-scanner-julz==0.2
zebra==0.1.0
